requirements.txt
pydantic==2.11.7
protobuf==6.32.0
google-api-python-client==2.179.0
python-dotenv==1.1.1
environs==14.3.0
pydantic_settings==2.10.1
pandas==2.3.1
asyncio==4.0.0
tenacity==9.1.2
httpx==0.28.1
